requirements.txt
ansicon==1.89.0
bcrypt==4.1.3
blessed==1.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
editor==1.6.6
idna==3.7
inquirer==3.2.4
jinxed==1.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
pandas==2.2.2
paramiko==3.4.0
pycparser==2.22
pyfiglet==1.0.2
Pygments==2.18.0
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
readchar==4.1.0
requests==2.32.3
rich==13.7.1
runs==1.2.2
six==1.16.0
sshtunnel==0.4.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xmod==1.8.1